feat(secret-approval): moved approval code to ee
