+++ backend/src/controllers/v1/index.ts
@@ -16,8 +16,6 @@ import * as workspaceController from "./workspaceController";
 import * as secretScanningController from "./secretScanningController";
 import * as webhookController from "./webhookController";
 import * as secretImpsController from "./secretImpsController";
-import * as secretApprovalPolicyController from "./secretApprovalPolicyController";
-import * as secretApprovalRequestController from "./secretApprovalRequestsController";
 export {
   authController,
   botController,
@@ -36,7 +34,5 @@ export {
   workspaceController,
   secretScanningController,
   webhookController,
-  secretImpsController,
-  secretApprovalPolicyController,
-  secretApprovalRequestController
+  secretImpsController
 };
