+++ backend/src/ee/validation/secretApproval.ts
