+++ backend/src/ee/validation/secretApprovalRequest.ts
