+++ backend/src/validation/index.ts
@@ -1,4 +1,3 @@
-export * from "./secretApproval";
 export * from "./user";
 export * from "./workspace";
 export * from "./bot";
